Add types using JSDoc to the `graphlib` file (#60)
* test: type `var conflicts` vars in `bk.test.js`

This fixes any potential type errors when calling the `addConflict`
function in our unit tests, after the changes from
dbe53cb (fix: Prototype Pollution (#54), 2025-10-21).

Fixes: dbe53cb81afd318d9d800bb21167374df262b8df

* test: import missing `beforeEach` and `afterEach`

Vitest adds these automatically, so it didn't cause any issues when
running the tests, but it did cause issues with TypeScript.

* test: type `var g` with `Graph`

This lets TypeScript check that the functions we call on this type are
valid.

* build: change `tsconfig.json` to type-check tests

Update the `tsconfig.json` file to also type-check the tests, checking
for any potential mistakes we make in TypeScript.

The new `tsconfig.build.json` file is then used for actually building
the types.

* fix(types): improve horizontalCompaction type

* feat(types): improve `Graph` documented types

I've manually gone through the JSDoc references within the
`src/graphlib/graph.js` file and added JSDoc comments when TypeScript
couldn't automatically find detect the types.

I've tried to keep the types as backwards compatible as possible.

There's also an MIT-licensed [`@types/graphlib`][1] package that we can
copy code from, but I didn't do that since I noticed issues with it that
were causing our unit tests to fail type-checking (e.g. not allowing
`number`s as params for lots of functions, even though the test code
uses it a lot).

[1]: http://npmjs.com/package/@types/graphlib

* docs: copy over graphlib documentation for Graph

Update the Graph documentation with a mixture of:

1. API documentation from the graphlib wiki:
   https://github.com/dagrejs/graphlib/wiki/API-Reference
2. Documentation from
   https://github.com/dagrejs/graphlib/blob/64b31ef3df0270c9bda85b45cb338481ce25618d/lib/graph.js

Both of these are from the graphlib MIT license, which we're
already complying with.